ToxicGas ult

+++ NoDesignersOnlyChaos/Assets/Scripts/Health.cs
@@ -54,7 +54,7 @@ public class Health : MonoBehaviour, IDamageable
         }
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         PlayTakeDamageSound();
         health -= damage;
